Top Firmware page fixed
