+++ src/pages/firmwareView/FirmwareView.module.scss
