+++ src/pages/firmwareView/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "FirmwareView",
+  "version": "0.0.0",
+  "private": true,
+  "main": "FirmwareView.js"
+}
