+++ src/pages/topFirmware/package.json
@@ -1,6 +1,6 @@
 {
   "name": "TopFirmware",
-  "version": "0.0.0",
-  "private": true,
-  "main": "TopFirmware.js"
+  "version": "1.0.0",
+  "main": "TopFirmware.js",
+  "license": "MIT"
 }
